Reworked individual sections of content with proper meta and dynamic nav
This refactors the policy pieces of content to match how we were doing prework so that we can compose a single path through all of the content for the students. That required moving that content out of the root into a new _sections directory to replace the existing _prework directory. This required changes to config and how we build the header as well. Some minor styling changes came along with this so that we can use proper semantic html.